refactor: create packages for all classes
