+++ src/main/java/com/example/expensesplitter/controller/GroupController.java
@@ -1,12 +1,11 @@
 package com.example.expensesplitter.controller;
 
 import com.example.expensesplitter.dto.request.group.GroupRequestDto;
-import com.example.expensesplitter.dto.response.SuccessResponseDto;
 import com.example.expensesplitter.dto.response.group.GroupResponseDto;
+import com.example.expensesplitter.dto.response.success.SuccessResponseDto;
 import com.example.expensesplitter.service.GroupService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,19 +24,19 @@ public class GroupController {
     public ResponseEntity<SuccessResponseDto> getGroups() {
         List<GroupResponseDto> groups = groupService.getGroups();
         return ResponseEntity.ok(SuccessResponseDto.<List<GroupResponseDto>>builder()
-                .message("Groups retrieved.")
-                .data(groups)
-                .build()
-        );
+                                                   .message("Groups retrieved.")
+                                                   .data(groups)
+                                                   .build()
+                                );
     }
 
     @GetMapping("/{groupId}")
     public ResponseEntity<SuccessResponseDto> getGroup(@PathVariable String groupId) {
         GroupResponseDto group = groupService.getGroupById(groupId);
         return ResponseEntity.ok(SuccessResponseDto.<GroupResponseDto>builder()
-                .message("Group retrieved successfully.")
-                .data(group)
-                .build());
+                                                   .message("Group retrieved successfully.")
+                                                   .data(group)
+                                                   .build());
     }
 
     @PostMapping
@@ -45,8 +44,8 @@ public class GroupController {
         GroupResponseDto createdGroup = groupService.createGroup(group);
 
         return ResponseEntity.ok(SuccessResponseDto.<GroupResponseDto>builder()
-                .message("Group created.")
-                .data(createdGroup)
-                .build());
+                                                   .message("Group created.")
+                                                   .data(createdGroup)
+                                                   .build());
     }
 }
