+++ src/main/java/com/example/expensesplitter/dto/response/success/SuccessResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.expensesplitter.dto.response;
+package com.example.expensesplitter.dto.response.success;
 
 import lombok.*;
