+++ src/main/java/com/example/expensesplitter/entity/Group.java
@@ -1,8 +1,8 @@
 package com.example.expensesplitter.entity;
 
-import com.example.expensesplitter.enums.Currency;
-import com.example.expensesplitter.enums.GroupStatus;
-import com.example.expensesplitter.enums.GroupType;
+import com.example.expensesplitter.enums.group.Currency;
+import com.example.expensesplitter.enums.group.GroupStatus;
+import com.example.expensesplitter.enums.group.GroupType;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
