+++ src/main/java/com/example/expensesplitter/security/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.example.expensesplitter.security;
+package com.example.expensesplitter.security.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
